Fix clippy warnings
This patch applies all clippy lints currently present on the latest
clippy master than are compatible with our oldstable clippy (only
exception is the `_else(||` stuff).

+++ alacritty/src/renderer/rects.rs
@@ -333,7 +333,7 @@ impl RectRenderer {
                     continue;
                 }
 
-                self.program.set_rect_kind(rect_kind as u8);
+                self.program.set_rect_kind(rect_kind);
 
                 // Upload accumulated undercurl vertices.
                 gl::BufferData(
